move lib contents to nimbus file
