+++ components/nimbus/src/tests/mod.rs
@@ -7,7 +7,7 @@ mod test_defaults;
 mod test_enrollment;
 mod test_enrollment_bw_compat;
 mod test_evaluator;
-mod test_lib;
+mod test_nimbus;
 mod test_lib_bw_compat;
 mod test_lib_schema_deserialization;
 mod test_persistence;
