+++ components/nimbus/src/tests/test_lib_schema_deserialization.rs
@@ -2,7 +2,8 @@
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::{error::Result, Deserialize, FeatureConfig, Serialize};
+use crate::{error::Result, FeatureConfig};
+use serde::{Deserialize, Serialize};
 use serde_json::{json, Map, Value};
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
