+++ components/nimbus/src/tests/test_nimbus.rs
@@ -10,9 +10,10 @@ use crate::{
     enrollment::{EnrolledReason, EnrollmentStatus, ExperimentEnrollment},
     error::Result,
     persistence::Database,
-    AppContext, AvailableRandomizationUnits, Experiment, NimbusClient, Path, StoreId,
+    AppContext, AvailableRandomizationUnits, Experiment, NimbusClient, persistence::StoreId,
     TargetingAttributes, DB_KEY_APP_VERSION, DB_KEY_UPDATE_DATE,
 };
+use std::path::Path;
 use chrono::{DateTime, Duration, Utc};
 use serde_json::json;
 use std::io::Write;
